refactor(expr): rearrange common_expression and common_function
