+++ src/query/expression/src/lib.rs
@@ -37,41 +37,31 @@ pub type Result<T> = std::result::Result<T, (crate::Span, String)>;
 #[allow(dead_code)]
 mod block;
 
-mod block_compact_thresholds;
-pub mod block_debug;
 pub mod converts;
 mod deserializations;
 mod evaluator;
 mod expression;
 mod function;
 mod kernels;
-pub mod large_number;
-mod memory;
-mod meta_info;
 mod property;
 mod register;
 pub mod schema;
-mod stream;
 pub mod type_check;
 pub mod types;
 pub mod utils;
 pub mod values;
 
-pub use meta_info::BlockMetaInfo;
-pub use meta_info::BlockMetaInfoPtr;
-pub use meta_info::BlockMetaInfos;
-pub use stream::SendableDataBlockStream;
-
+pub use crate::block::BlockMetaInfo;
+pub use crate::block::BlockMetaInfoPtr;
 pub use crate::block::*;
-pub use crate::block_compact_thresholds::BlockCompactThresholds;
 pub use crate::deserializations::*;
 pub use crate::evaluator::*;
 pub use crate::expression::*;
 pub use crate::function::*;
 pub use crate::kernels::*;
-pub use crate::memory::InMemoryData;
 pub use crate::property::*;
 pub use crate::register::*;
 pub use crate::schema::*;
+pub use crate::utils::block_compact_thresholds::BlockCompactThresholds;
 pub use crate::utils::*;
 pub use crate::values::*;
