--- src/query/expression/src/meta_info.rs
@@ -1,50 +1,0 @@
-// Copyright 2022 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use std::any::Any;
-use std::fmt::Debug;
-
-#[typetag::serde(tag = "type")]
-pub trait BlockMetaInfo: Debug + Send + Sync {
-    fn as_any(&self) -> &dyn Any;
-    fn as_mut_any(&mut self) -> &mut dyn Any;
-
-    #[allow(clippy::borrowed_box)]
-    fn equals(&self, info: &Box<dyn BlockMetaInfo>) -> bool;
-
-    fn clone_self(&self) -> Box<dyn BlockMetaInfo>;
-}
-
-impl Eq for Box<dyn BlockMetaInfo> {}
-
-impl PartialEq for Box<dyn BlockMetaInfo> {
-    fn eq(&self, other: &Self) -> bool {
-        let this_type_id = self.as_any().type_id();
-        let other_type_id = other.as_any().type_id();
-
-        match this_type_id == other_type_id {
-            true => self.equals(other),
-            false => false,
-        }
-    }
-}
-
-impl Clone for Box<dyn BlockMetaInfo> {
-    fn clone(&self) -> Self {
-        self.clone_self()
-    }
-}
-
-pub type BlockMetaInfoPtr = Box<dyn BlockMetaInfo>;
-pub type BlockMetaInfos = Vec<BlockMetaInfoPtr>;
