+++ src/query/expression/src/type_check.rs
@@ -189,7 +189,7 @@ pub fn check_function<Index: ColumnIndex>(
     let candidates = fn_registry.search_candidates(name, params, args);
 
     if candidates.is_empty() && !fn_registry.contains(name) {
-        return Err((span, format!("function `{name}` doesn't exist")));
+        return Err((span, format!("function `{name}` does not exist")));
     }
 
     let additional_rules = fn_registry
