+++ src/query/functions/src/scalars/other.rs
@@ -12,33 +12,37 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::io::Write;
 use std::net::Ipv4Addr;
 use std::sync::Arc;
 use std::time::Duration;
 
 use common_base::base::convert_byte_size;
 use common_base::base::convert_number_size;
+use common_base::base::uuid::Uuid;
+use common_expression::error_to_null;
 use common_expression::types::boolean::BooleanDomain;
 use common_expression::types::nullable::NullableColumn;
 use common_expression::types::number::Float64Type;
+use common_expression::types::number::Int64Type;
 use common_expression::types::number::UInt32Type;
 use common_expression::types::number::UInt8Type;
 use common_expression::types::number::F64;
+use common_expression::types::string::StringColumn;
 use common_expression::types::ArgType;
 use common_expression::types::DataType;
 use common_expression::types::DateType;
 use common_expression::types::GenericType;
 use common_expression::types::NullType;
 use common_expression::types::NullableType;
-use common_expression::types::NumberDataType;
 use common_expression::types::NumberType;
+use common_expression::types::SimpleDomain;
 use common_expression::types::StringType;
 use common_expression::types::TimestampType;
 use common_expression::types::ValueType;
-use common_expression::types::ALL_NUMERICS_TYPES;
 use common_expression::vectorize_with_builder_1_arg;
-use common_expression::with_number_mapped_type;
 use common_expression::Domain;
+use common_expression::EvalContext;
 use common_expression::Function;
 use common_expression::FunctionDomain;
 use common_expression::FunctionProperty;
@@ -48,6 +52,8 @@ use common_expression::Scalar;
 use common_expression::Value;
 use common_expression::ValueRef;
 use ordered_float::OrderedFloat;
+use rand::Rng;
+use rand::SeedableRng;
 
 pub fn register(registry: &mut FunctionRegistry) {
     registry.register_aliases("inet_aton", &["ipv4_string_to_num"]);
@@ -55,6 +61,11 @@ pub fn register(registry: &mut FunctionRegistry) {
     registry.register_aliases("inet_ntoa", &["ipv4_num_to_string"]);
     registry.register_aliases("try_inet_ntoa", &["try_ipv4_num_to_string"]);
     registry.register_aliases("assume_not_null", &["remove_nullable"]);
+    registry.register_aliases("gen_random_uuid", &["uuid"]);
+
+    register_inet_aton(registry);
+    register_inet_ntoa(registry);
+    register_run_diff(registry);
 
     registry.register_passthrough_nullable_1_arg::<Float64Type, StringType, _, _>(
         "humanize_size",
@@ -104,6 +115,39 @@ pub fn register(registry: &mut FunctionRegistry) {
                 }
             }
         }),
+    );
+
+    registry.register_0_arg_core::<NumberType<F64>, _, _>(
+        "rand",
+        FunctionProperty::default().non_deterministic(),
+        || {
+            FunctionDomain::Domain(SimpleDomain {
+                min: OrderedFloat(0.0),
+                max: OrderedFloat(1.0),
+            })
+        },
+        |ctx| {
+            let mut rng = rand::rngs::SmallRng::from_entropy();
+            let rand_nums = (0..ctx.num_rows)
+                .map(|_| rng.gen::<F64>())
+                .collect::<Vec<_>>();
+            Value::Column(rand_nums.into())
+        },
+    );
+
+    registry.register_1_arg::<NumberType<u64>, NumberType<F64>, _, _>(
+        "rand",
+        FunctionProperty::default().non_deterministic(),
+        |_| {
+            FunctionDomain::Domain(SimpleDomain {
+                min: OrderedFloat(0.0),
+                max: OrderedFloat(1.0),
+            })
+        },
+        |val, _| {
+            let mut rng = rand::rngs::SmallRng::seed_from_u64(val);
+            rng.gen::<F64>()
+        },
     );
 
     registry.register_1_arg_core::<GenericType<0>, StringType, _, _>(
@@ -163,10 +207,46 @@ pub fn register(registry: &mut FunctionRegistry) {
             |val, _| val.to_owned(),
         );
 
+    registry.register_0_arg_core::<StringType, _, _>(
+        "gen_random_uuid",
+        FunctionProperty::default(),
+        || FunctionDomain::Full,
+        |ctx| {
+            let mut values: Vec<u8> = Vec::with_capacity(ctx.num_rows * 36);
+            let mut offsets: Vec<u64> = Vec::with_capacity(ctx.num_rows);
+            offsets.push(0);
+
+            for _ in 0..ctx.num_rows {
+                let value = Uuid::new_v4();
+                offsets.push(offsets.last().unwrap() + 36u64);
+                write!(&mut values, "{:x}", value).unwrap();
+            }
+
+            let col = StringColumn {
+                data: values.into(),
+                offsets: offsets.into(),
+            };
+            Value::Column(col)
+        },
+    );
+}
+
+fn register_inet_aton(registry: &mut FunctionRegistry) {
     registry.register_passthrough_nullable_1_arg::<StringType, UInt32Type, _, _>(
         "inet_aton",
         FunctionProperty::default(),
         |_| FunctionDomain::MayThrow,
+        eval_inet_aton,
+    );
+
+    registry.register_combine_nullable_1_arg::<StringType, UInt32Type, _, _>(
+        "try_inet_aton",
+        FunctionProperty::default(),
+        |_| FunctionDomain::Full,
+        error_to_null(eval_inet_aton),
+    );
+
+    fn eval_inet_aton(val: ValueRef<StringType>, ctx: &mut EvalContext) -> Value<UInt32Type> {
         vectorize_with_builder_1_arg::<StringType, UInt32Type>(|v, output, ctx| {
             let addr_str = String::from_utf8_lossy(v);
             match addr_str.parse::<Ipv4Addr>() {
@@ -179,76 +259,43 @@ pub fn register(registry: &mut FunctionRegistry) {
                     output.push(0);
                 }
             }
-        }),
+        })(val, ctx)
+    }
+}
+
+fn register_inet_ntoa(registry: &mut FunctionRegistry) {
+    registry.register_passthrough_nullable_1_arg::<Int64Type, StringType, _, _>(
+        "inet_ntoa",
+        FunctionProperty::default(),
+        |_| FunctionDomain::MayThrow,
+        eval_inet_ntoa,
     );
 
-    registry.register_combine_nullable_1_arg::<StringType, UInt32Type, _, _>(
-        "try_inet_aton",
+    registry.register_combine_nullable_1_arg::<Int64Type, StringType, _, _>(
+        "try_inet_ntoa",
         FunctionProperty::default(),
         |_| FunctionDomain::Full,
-        vectorize_with_builder_1_arg::<StringType, NullableType<UInt32Type>>(|v, output, _| {
-            let addr_str = String::from_utf8_lossy(v);
-            match addr_str.parse::<Ipv4Addr>() {
-                Ok(addr) => {
-                    let addr_binary = u32::from(addr);
-                    output.push(addr_binary);
-                }
-                Err(_) => output.push_null(),
-            }
-        }),
+        error_to_null(eval_inet_ntoa),
     );
 
-    for num_ty in ALL_NUMERICS_TYPES {
-        with_number_mapped_type!(|NUM_TYPE| match num_ty {
-            NumberDataType::NUM_TYPE => {
-                registry
-                    .register_passthrough_nullable_1_arg::<NumberType<NUM_TYPE>, StringType, _, _>(
-                        "inet_ntoa",
-                        FunctionProperty::default(),
-                        |_| FunctionDomain::MayThrow,
-                        vectorize_with_builder_1_arg::<NumberType<NUM_TYPE>, StringType>(
-                            |v, output, ctx| match num_traits::cast::cast::<NUM_TYPE, u32>(v) {
-                                Some(val) => {
-                                    let addr_str = Ipv4Addr::from(val.to_be_bytes()).to_string();
-                                    output.put_str(&addr_str);
-                                    output.commit_row();
-                                }
-                                None => {
-                                    ctx.set_error(
-                                        output.len(),
-                                        format!("Failed to parse '{}' into a IPV4 address", v),
-                                    );
-                                    output.commit_row();
-                                }
-                            },
-                        ),
+    fn eval_inet_ntoa(val: ValueRef<Int64Type>, ctx: &mut EvalContext) -> Value<StringType> {
+        vectorize_with_builder_1_arg::<Int64Type, StringType>(|val, output, ctx| {
+            match num_traits::cast::cast::<i64, u32>(val) {
+                Some(val) => {
+                    let addr_str = Ipv4Addr::from(val.to_be_bytes()).to_string();
+                    output.put_str(&addr_str);
+                    output.commit_row();
+                }
+                None => {
+                    ctx.set_error(
+                        output.len(),
+                        format!("Failed to parse '{}' into a IPV4 address", val),
                     );
+                    output.commit_row();
+                }
             }
-        });
-
-        with_number_mapped_type!(|NUM_TYPE| match num_ty {
-            NumberDataType::NUM_TYPE => {
-                registry.register_combine_nullable_1_arg::<NumberType<NUM_TYPE>, StringType, _, _>(
-                    "try_inet_ntoa",
-                    FunctionProperty::default(),
-                    |_| FunctionDomain::Full,
-                    vectorize_with_builder_1_arg::<NumberType<NUM_TYPE>, NullableType<StringType>>(
-                        |v, output, _| match num_traits::cast::cast::<NUM_TYPE, u32>(v) {
-                            Some(val) => {
-                                let addr_str = Ipv4Addr::from(val.to_be_bytes()).to_string();
-                                output.validity.push(true);
-                                output.builder.put_str(&addr_str);
-                                output.builder.commit_row();
-                            }
-                            None => output.push_null(),
-                        },
-                    ),
-                );
-            }
-        });
+        })(val, ctx)
     }
-
-    register_run_diff(registry);
 }
 
 macro_rules! register_simple_domain_type_run_diff {
