+++ src/query/functions/tests/it/scalars/boolean.rs
@@ -64,16 +64,6 @@ fn test_and(file: &mut impl Write) {
     run_ast(file, "(0 > 1) AND (a < 1)", one_null_column().as_slice()); // false        AND NULL(true)
     run_ast(file, "(0 < 1) AND (a > 1)", one_null_column().as_slice()); // true         AND NULL(false)
     run_ast(file, "(0 < 1) AND (a < 1)", one_null_column().as_slice()); // true         AND NULL(true)
-
-    // Constant function call, even though it may throw, should not stop constant folding
-    run_ast(file, "const_false AND CAST('1000' AS UINT32) = 1000", &[(
-        "const_false",
-        BooleanType::from_data(vec![false]),
-    )]);
-    run_ast(file, "false AND CAST(str AS UINT32) = 1000", &[(
-        "str",
-        StringType::from_data(vec!["1000"]),
-    )]);
 }
 
 fn test_not(file: &mut impl Write) {
