+++ src/query/functions/tests/it/scalars/testdata/function_list.txt
@@ -1004,26 +1004,8 @@ humanize_size(Float64) :: String
 humanize_size(Float64 NULL) :: String NULL
 inet_aton(String) :: UInt32
 inet_aton(String NULL) :: UInt32 NULL
-inet_ntoa(UInt8) :: String
-inet_ntoa(UInt8 NULL) :: String NULL
-inet_ntoa(UInt16) :: String
-inet_ntoa(UInt16 NULL) :: String NULL
-inet_ntoa(UInt32) :: String
-inet_ntoa(UInt32 NULL) :: String NULL
-inet_ntoa(UInt64) :: String
-inet_ntoa(UInt64 NULL) :: String NULL
-inet_ntoa(Int8) :: String
-inet_ntoa(Int8 NULL) :: String NULL
-inet_ntoa(Int16) :: String
-inet_ntoa(Int16 NULL) :: String NULL
-inet_ntoa(Int32) :: String
-inet_ntoa(Int32 NULL) :: String NULL
 inet_ntoa(Int64) :: String
 inet_ntoa(Int64 NULL) :: String NULL
-inet_ntoa(Float32) :: String
-inet_ntoa(Float32 NULL) :: String NULL
-inet_ntoa(Float64) :: String
-inet_ntoa(Float64 NULL) :: String NULL
 insert(String, Int64, Int64, String) :: String
 insert(String NULL, Int64 NULL, Int64 NULL, String NULL) :: String NULL
 instr(String, String) :: UInt64
@@ -2780,26 +2762,8 @@ truncate(Float64, Int64) :: Float64
 truncate(Float64 NULL, Int64 NULL) :: Float64 NULL
 try_inet_aton(String) :: UInt32 NULL
 try_inet_aton(String NULL) :: UInt32 NULL
-try_inet_ntoa(UInt8) :: String NULL
-try_inet_ntoa(UInt8 NULL) :: String NULL
-try_inet_ntoa(UInt16) :: String NULL
-try_inet_ntoa(UInt16 NULL) :: String NULL
-try_inet_ntoa(UInt32) :: String NULL
-try_inet_ntoa(UInt32 NULL) :: String NULL
-try_inet_ntoa(UInt64) :: String NULL
-try_inet_ntoa(UInt64 NULL) :: String NULL
-try_inet_ntoa(Int8) :: String NULL
-try_inet_ntoa(Int8 NULL) :: String NULL
-try_inet_ntoa(Int16) :: String NULL
-try_inet_ntoa(Int16 NULL) :: String NULL
-try_inet_ntoa(Int32) :: String NULL
-try_inet_ntoa(Int32 NULL) :: String NULL
 try_inet_ntoa(Int64) :: String NULL
 try_inet_ntoa(Int64 NULL) :: String NULL
-try_inet_ntoa(Float32) :: String NULL
-try_inet_ntoa(Float32 NULL) :: String NULL
-try_inet_ntoa(Float64) :: String NULL
-try_inet_ntoa(Float64 NULL) :: String NULL
 try_parse_json(String) :: Variant NULL
 try_parse_json(String NULL) :: Variant NULL
 try_parse_json(Boolean) :: Variant NULL
