+++ src/query/functions/tests/it/scalars/testdata/other.txt
@@ -226,7 +226,7 @@ output         : 167773449_u32
 
 ast            : inet_ntoa(16909060)
 raw expr       : inet_ntoa(16909060_u32)
-checked expr   : inet_ntoa<UInt32>(16909060_u32)
+checked expr   : inet_ntoa<Int64>(CAST(16909060_u32 AS Int64))
 optimized expr : "1.2.3.4"
 output type    : String
 output domain  : {"1.2.3.4"..="1.2.3.4"}
@@ -235,7 +235,7 @@ output         : "1.2.3.4"
 
 ast            : try_inet_ntoa(121211111111111)
 raw expr       : try_inet_ntoa(121211111111111_u64)
-checked expr   : try_inet_ntoa<UInt64>(121211111111111_u64)
+checked expr   : try_inet_ntoa<Int64>(CAST(121211111111111_u64 AS Int64))
 optimized expr : NULL
 output type    : String NULL
 output domain  : {NULL}
