+++ src/query/service/src/pipelines/processors/transforms/group_by/mod.rs
@@ -18,6 +18,7 @@ mod aggregator_keys_iter;
 mod aggregator_polymorphic_keys;
 mod aggregator_state;
 mod aggregator_state_entity;
+mod large_number;
 
 pub use aggregator_groups_builder::GroupColumnsBuilder;
 pub use aggregator_keys_builder::KeysColumnBuilder;
