+++ src/query/sql/src/executor/format.rs
@@ -16,7 +16,6 @@ use common_ast::ast::FormatTreeNode;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_expression::ConstantFolder;
-use common_expression::Domain;
 use common_expression::FunctionContext;
 use common_functions::scalars::BUILTIN_FUNCTIONS;
 use itertools::Itertools;
@@ -74,7 +73,6 @@ fn table_scan_to_format_tree(
     if plan.table_index == DUMMY_TABLE_INDEX {
         return Ok(FormatTreeNode::new("DummyTableScan".to_string()));
     }
-    let func_ctx = FunctionContext::default();
     let table = metadata.read().table(plan.table_index).clone();
     let table_name = format!("{}.{}.{}", table.catalog(), table.database(), table.name());
     let filters = plan
@@ -87,16 +85,8 @@ fn table_scan_to_format_tree(
                 .iter()
                 .map(|f| {
                     let expr = f.as_expr(&BUILTIN_FUNCTIONS).unwrap();
-                    let input_domains = expr
-                        .column_refs()
-                        .into_iter()
-                        .map(|(name, ty)| {
-                            let domain = Domain::full(&ty);
-                            (name, domain)
-                        })
-                        .collect();
-                    let folder = ConstantFolder::new(input_domains, func_ctx, &BUILTIN_FUNCTIONS);
-                    let (new_expr, _) = folder.fold(&expr);
+                    let (new_expr, _) =
+                        ConstantFolder::fold(&expr, FunctionContext::default(), &BUILTIN_FUNCTIONS);
                     new_expr.sql_display()
                 })
                 .collect::<Vec<_>>()
